Generate ChartHeader based on MATContext
